@@ -1,25 +1,14 @@
-// Top-level build file where you can add configuration options common to all sub-projects/modules.
-
 buildscript {
-    ext.kotlin_version = '1.3.30'
+    ext.kotlin_version = '1.6.10'
+
     repositories {
         google()
-        jcenter()
+        mavenCentral()
         
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.4.0'
+        classpath 'com.android.tools.build:gradle:7.0.4'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
-        // NOTE: Do not place your application dependencies here; they belong
-        // in the individual module build.gradle files
-    }
-}
-
-allprojects {
-    repositories {
-        google()
-        jcenter()
-        
     }
 }
 
@@ -27,10 +16,3 @@ task clean(type: Delete) {
     delete rootProject.buildDir
 }
 
-ext {
-    android_version = '28.0.0'
-    constraint_layout_version = '1.1.3'
-    junit_version = '4.12'
-    test_runner_version = '1.0.2'
-    test_espresso_version = '3.0.2'
-}
