@@ -2,7 +2,7 @@ package com.zero.sample
 
 import android.content.Context
 import android.content.res.ColorStateList
-import android.support.annotation.ColorInt
+import androidx.annotation.ColorInt
 import android.util.AttributeSet
 import android.view.LayoutInflater
 import android.view.View
