@@ -1,7 +1,7 @@
 package com.zero.sample
 
 import android.annotation.SuppressLint
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.view.View
 import com.zero.chartview.BuildConfig
